1, Add RenewBackupPlan DBS interface.

+++ services/dbs/struct_restore_task_detail.go
@@ -26,12 +26,12 @@ type RestoreTaskDetail struct {
 	DestinationEndpointUserName     string `json:"DestinationEndpointUserName" xml:"DestinationEndpointUserName"`
 	DestinationEndpointOracleSID    string `json:"DestinationEndpointOracleSID" xml:"DestinationEndpointOracleSID"`
 	RestoreObjects                  string `json:"RestoreObjects" xml:"RestoreObjects"`
-	BackupGatewayId                 int    `json:"BackupGatewayId" xml:"BackupGatewayId"`
+	BackupGatewayId                 int64  `json:"BackupGatewayId" xml:"BackupGatewayId"`
 	RestoreDir                      string `json:"RestoreDir" xml:"RestoreDir"`
 	RestoreTaskName                 string `json:"RestoreTaskName" xml:"RestoreTaskName"`
 	BackupSetId                     string `json:"BackupSetId" xml:"BackupSetId"`
-	RestoreTime                     int    `json:"RestoreTime" xml:"RestoreTime"`
-	RestoreTaskCreateTime           int    `json:"RestoreTaskCreateTime" xml:"RestoreTaskCreateTime"`
+	RestoreTime                     int64  `json:"RestoreTime" xml:"RestoreTime"`
+	RestoreTaskCreateTime           int64  `json:"RestoreTaskCreateTime" xml:"RestoreTaskCreateTime"`
 	RestoreStatus                   string `json:"RestoreStatus" xml:"RestoreStatus"`
 	RestoreTaskId                   string `json:"RestoreTaskId" xml:"RestoreTaskId"`
 	FullStruforeRestoreProgress     int    `json:"FullStruforeRestoreProgress" xml:"FullStruforeRestoreProgress"`
